# backend/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "ICTSC Score Server API",
    "version": "1.0.0",
    "description": "ICTSCの競技者・運営向けREST API。問題、Sections設定、お知らせはGitHubを正本としてcommit単位のsnapshotで中継し、それ以外の競技状態はデータベースで管理する。Cookie認証を使う全mutationは同一Originを必須とする。GitHub Actions OIDCまたはSState固定Bearerを使うmachine-to-machine endpointにはOriginを要求しない。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同一オリジン"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/content/actions/refresh": {
      "post": {
        "tags": ["Admin Content"],
        "summary": "指定commitからコンテンツsnapshotを更新する",
        "description": "Admin sessionまたはGitHub Actions OIDCで実行する。同一commitは冪等に200を返す。OIDC利用時はissuer、audience、repository_id、ref、job_workflow_refを検証する。現在より古いworkflow実行によるactive snapshotの巻き戻しは409を返す。manifestと参照Markdownを全て検証し、成功時だけactive snapshotを原子的に切り替える。",
        "operationId": "refreshAdminContent",
        "security": [
          { "AdminSession": [] },
          { "GitHubActionsOIDC": [] }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshContentRequest" } } }
        },
        "responses": {
          "200": {
            "description": "snapshotを有効化した、または同じcommitが既に有効だった",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContentStatusResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "502": { "$ref": "#/components/responses/BadGateway" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ContestantSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "user-session",
        "description": "Discord認証と競技者登録の完了後に3日間発行されるopaqueなHttpOnly Cookie。本番ではSecure、SameSite=Strict。"
      },
      "SignupSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "signup-session",
        "description": "Discord認証済みだが競技者登録前の状態を10分間表すopaqueなHttpOnly Cookie。本番ではSecure、SameSite=Strict。"
      },
      "OAuthSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oauth2-session",
        "description": "競技者Discord OAuthのstate、PKCE verifier、遷移先を10分間保持するopaqueなHttpOnly Cookie。本番ではSecure、SameSite=Lax。"
      },
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin-session",
        "description": "Discord guild roleの検証後に8時間発行されるopaqueなAdmin HttpOnly Cookie。本番ではSecure、SameSite=Strict。"
      },
      "AdminOAuthSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin-oauth2-session",
        "description": "Admin Discord OAuthのstate、PKCE verifier、遷移先を10分間保持するopaqueなHttpOnly Cookie。本番ではSecure、SameSite=Lax。"
      },
      "GitHubActionsOIDC": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "GitHub Actions OIDC JWT",
        "description": "content refresh専用。GitHub issuer、設定済みaudience、repository_id、ref、job_workflow_refを全て検証する短期OIDC token。"
      },
      "SStateCallbackBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Opaque",
...
    "responses": {
...
      "Unauthorized": {
        "description": "認証情報がない、期限切れ、または不正",
        "x-error-codes": ["authentication_required", "invalid_session", "oauth_state_invalid", "invalid_machine_token"],
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
      },
      "Forbidden": {
        "description": "現在のユーザー状態または権限では実行できない",
        "x-error-codes": ["permission_denied", "origin_forbidden", "admin_role_required", "guild_membership_required"],
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
      },
...
      "Conflict": {
        "description": "重複、期間外操作、状態遷移違反などのドメイン競合",
        "x-error-codes": ["conflict", "team_code_conflict", "invitation_expired", "invitation_already_used", "team_full", "contestant_already_registered", "submission_closed", "deployment_not_allowed", "deployment_in_progress", "invalid_deployment_transition", "duplicate_event_mismatch", "content_rollback_rejected", "content_refresh_in_progress"],
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
      },
      "ValidationError": {
        "description": "入力値またはリクエスト形式が不正",
        "x-error-codes": ["validation_error", "content_invalid"],
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
      },
...
      "InternalServerError": {
        "description": "内部エラー",
        "x-error-codes": ["internal_error"],
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
      },
...
      "BadGateway": {
        "description": "GitHubまたはSStateに接続できない、または初回content snapshotが存在しない",
        "x-error-codes": ["upstream_unavailable", "content_not_available"],
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
      }
    },
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": false,
        "description": "RFC 9457 Problem Details",
        "required": ["type", "title", "status", "detail", "code"],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference",
            "default": "about:blank"
          },
          "title": { "type": "string", "minLength": 1 },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "detail": { "type": "string", "minLength": 1 },
          "instance": { "type": ["string", "null"], "format": "uri-reference" },
          "code": { "$ref": "#/components/schemas/ErrorCode" },
          "errors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ErrorDetail" },
            "default": []
          }
        },
        "examples": [
          {
            "type": "https://score.example.invalid/problems/validation-error",
            "title": "Validation failed",
            "status": 422,
            "detail": "One or more request fields are invalid.",
            "instance": "/api/v1/auth/signup",
            "code": "validation_error",
            "errors": [
              { "location": "body.invitation_code", "message": "must not be empty" }
            ]
          }
        ]
      },
      "ErrorCode": {
        "type": "string",
        "description": "クライアントが分岐に利用できる後方互換なRFC 9457 code。各responseのx-error-codesがそのstatusで返り得る値を示す。",
        "enum": [
          "authentication_required",
          "invalid_session",
          "oauth_state_invalid",
          "invalid_machine_token",
          "permission_denied",
          "origin_forbidden",
          "admin_role_required",
          "guild_membership_required",
          "resource_not_found",
          "content_not_found",
          "validation_error",
          "conflict",
          "team_code_conflict",
          "invitation_expired",
          "invitation_already_used",
          "team_full",
          "contestant_already_registered",
          "answer_rate_limited",
          "submission_closed",
          "deployment_not_allowed",
          "deployment_in_progress",
          "invalid_deployment_transition",
          "duplicate_event_mismatch",
          "content_rollback_rejected",
          "content_refresh_in_progress",
          "upstream_unavailable",
          "content_not_available",
          "content_invalid",
          "internal_error"
        ]
      },
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": ["location", "message"],
        "properties": {
          "location": { "type": "string", "minLength": 1 },
          "message": { "type": "string", "minLength": 1 },
          "value": {}
        }
      },
...
      "ContentCommit": {
        "type": "string",
        "description": "検証済みcontent snapshotを一意に指すGit commit object ID。",
        "minLength": 40,
        "maxLength": 64,
        "pattern": "^[0-9a-f]{40,64}$"
      },
      "ContentRefreshState": {
        "type": "string",
        "enum": ["NEVER", "REFRESHING", "SUCCEEDED", "FAILED"]
      },
      "ContentStatus": {
        "type": "object",
        "additionalProperties": false,
        "description": "active_commitがnullなのは一度もsnapshotを有効化していない場合だけ。FAILEDでもactive_commitがあれば最終正常版を配信し続ける。",
        "required": ["state", "active_commit", "activated_at", "source_ref", "last_attempt_commit", "last_attempt_at", "last_error", "serving_last_known_good"],
        "properties": {
          "state": { "$ref": "#/components/schemas/ContentRefreshState" },
          "active_commit": {
            "oneOf": [
              { "$ref": "#/components/schemas/ContentCommit" },
              { "type": "null" }
            ]
          },
          "activated_at": { "type": ["string", "null"], "format": "date-time" },
          "source_ref": { "type": ["string", "null"], "maxLength": 255 },
          "last_attempt_commit": {
            "oneOf": [
              { "$ref": "#/components/schemas/ContentCommit" },
              { "type": "null" }
            ]
          },
          "last_attempt_at": { "type": ["string", "null"], "format": "date-time" },
          "last_error": { "type": ["string", "null"], "maxLength": 4000 },
          "serving_last_known_good": { "type": "boolean" }
        }
      },
      "ContentStatusResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["content"],
        "properties": {
          "content": { "$ref": "#/components/schemas/ContentStatus" }
        }
      },
      "RefreshContentRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["commit"],
        "properties": {
          "commit": { "$ref": "#/components/schemas/ContentCommit" }
        }
      },
...
    }
  }
}